Move classes to new package.
