+++ java/src/main/java/com/powsybl/dataframe/network/modifications/NetworkModifications.java
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl.dataframe.network.adders;
+package com.powsybl.dataframe.network.modifications;
 
 import com.powsybl.commons.PowsyblException;
 import com.powsybl.commons.reporter.ReporterModel;
